Modularise WalletDB and Manifest
