+++ src/libxrpl/server/Vacuum.cpp
@@ -1,4 +1,4 @@
-#include <xrpld/app/rdb/Vacuum.h>
+#include <xrpl/server/Vacuum.h>
 
 #include <boost/format.hpp>
