+++ src/xrpld/app/rdb/PeerFinder.h
@@ -1,8 +1,9 @@
 #pragma once
 
 #include <xrpld/core/Config.h>
-#include <xrpld/core/DatabaseCon.h>
 #include <xrpld/peerfinder/detail/Store.h>
+
+#include <xrpl/rdb/DatabaseCon.h>
 
 namespace xrpl {
